scripts/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.3.0
ccxt==1.60.58
python-dotenv==0.19.0
requests==2.26.0